+++ Plone/src/docpool.base/docpool/base/upgrades.py
@@ -422,3 +422,39 @@ def to_1008_remove_irix(context=None):
         api.content.delete(obj)
     portal_setup = api.portal.get_tool('portal_setup')
     loadMigrationProfile(portal_setup, 'profile-docpool.base:ticket_3954_remove_irix')
+
+
+def to_1008(context=None):
+    portal_setup = api.portal.get_tool('portal_setup')
+    log.info('Upgrading to 1008: adding report year index')
+    loadMigrationProfile(portal_setup, 'profile-docpool.base:to_1008')
+
+
+def to_1008_index_report_year(context=None):
+    log.info(u'Reindexing rei reports.')
+    brains = api.content.find(portal_type='DPDocument', dp_type='reireport')
+    log.info(u'Found {0} rei reports to reindex'.format(len(brains)))
+    for brain in brains:
+        obj = brain.getObject()
+        obj.reindexObject(idxs=['report_year'])
+    log.info('Done.')
+
+
+def to_1008_fix_unicode_indexes(context=None):
+    # Rebuild indexes with unicode values that fail in py2. See #4084
+    catalog = api.portal.get_tool('portal_catalog')
+    log.info('Rebuild index Origins ...')
+    catalog.manage_clearIndex(ids=['Origins'])
+    catalog.manage_reindexIndex(ids=['Origins'])
+    log.info('Rebuild index category ...')
+    catalog.manage_clearIndex(ids=['category'])
+    catalog.manage_reindexIndex(ids=['category'])
+    log.info('Done.')
+
+
+def to_1008_install_z3ctable(context=None):
+    portal = api.portal.get()
+    installer = get_installer(portal)
+    if not installer.is_product_installed('collective.eeafaceted.z3ctable'):
+        installer.install_product('collective.eeafaceted.z3ctable')
+        log.info(u'collective.eeafaceted.z3ctable installed')