+++ Plone/src/docpool.doksys/src/docpool/doksys/interfaces.py
@@ -3,6 +3,5 @@
 
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
-
 class IDocpoolDoksysLayer(IDefaultBrowserLayer):
     """Marker interface that defines a browser layer."""