+++ Plone/src/docpool.rei/docpool/rei/indexers.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+from docpool.base.content.dpdocument import IDPDocument
+from docpool.rei.config import REI_APP
+from plone.indexer import indexer
+
+
+@indexer(IDPDocument)
+def report_year(obj):
+    try:
+        return str(obj.doc_extension(REI_APP).Year)
+    except TypeError:
+        pass
+    return ''