+++ Plone/src/docpool.rei/docpool/rei/interfaces.py
@@ -1,1 +1,7 @@
 # -*- coding: utf-8 -*-
+from zope.interface import Interface
+
+
+class IReiMarker(Interface):
+    """Marker interface that defines a Zope 3 browser layer.
+    """