+++ Plone/src/docpool.theme/docpool/theme/webpack_resources/src/docpooltheme/faceted-z3ctable.js
@@ -1,0 +1,1 @@
+import('faceted-z3ctable');